Allow network_mode to be used in a step
With network mode we can reuse a network defined by another container (or the host itself), like a container connected
to a VPN. This is useful for the case that you should connect through a VPN to perform a deploy.

Because network_mode use an already created network, we can't use services on this build step. All custom networks
are disabled if network_mode is present.